numerous fixes
