+++ src/client_creds.rs
@@ -8,27 +8,6 @@ pub struct ClientCredentialsSpotify {
     pub(in crate) http: HttpClient,
 }
 
-impl ClientCredentialsSpotify {
-    pub fn new(creds: Credentials) -> Self {
-        ClientCredentialsSpotify {
-            creds,
-            ..Default::default()
-        }
-    }
-
-    pub fn with_config(creds: Credentials, config: Config) -> Self {
-        ClientCredentialsSpotify {
-            creds,
-            config,
-            ..Default::default()
-        }
-    }
-
-    pub fn request_token(&mut self) {
-        todo!()
-    }
-}
-
 // This could even use a macro
 impl BaseClient for ClientCredentialsSpotify {
     fn get_http(&self) -> &HttpClient {
@@ -45,5 +24,42 @@ impl BaseClient for ClientCredentialsSpotify {
 
     fn get_config(&self) -> &Config {
         &self.config
+    }
+}
+
+impl ClientCredentialsSpotify {
+    pub fn new(creds: Credentials) -> Self {
+        ClientCredentialsSpotify {
+            creds,
+            ..Default::default()
+        }
+    }
+
+    pub fn with_config(creds: Credentials, config: Config) -> Self {
+        ClientCredentialsSpotify {
+            creds,
+            config,
+            ..Default::default()
+        }
+    }
+
+    /// Obtains the client access token for the app without saving it into the
+    /// cache file. The resulting token is saved internally.
+    #[maybe_async]
+    pub async fn request_client_token_without_cache(&mut self) -> ClientResult<()> {
+        let mut data = Form::new();
+        data.insert(headers::GRANT_TYPE, headers::GRANT_CLIENT_CREDS);
+
+        self.token = Some(self.fetch_access_token(&data).await?);
+
+        Ok(())
+    }
+
+    /// The same as `request_client_token_without_cache`, but saves the token
+    /// into the cache file if possible.
+    #[maybe_async]
+    pub async fn request_client_token(&mut self) -> ClientResult<()> {
+        self.request_client_token_without_cache().await?;
+        self.write_token_cache()
     }
 }
