+++ src/code_auth.rs
@@ -1,11 +1,48 @@
-use crate::{http::HttpClient, prelude::*, Config, Credentials, OAuth, Token};
+use crate::{
+    auth_urls,
+    endpoints::{BaseClient, OAuthClient},
+    headers,
+    http::{Form, HttpClient},
+    ClientResult, Config, Credentials, OAuth, Token,
+};
+
+use std::collections::HashMap;
+
+use maybe_async::maybe_async;
+use url::Url;
 
 #[derive(Clone, Debug, Default)]
 pub struct CodeAuthSpotify {
     creds: Credentials,
     oauth: OAuth,
-    tok: Option<Token>,
+    config: Config,
+    token: Option<Token>,
     http: HttpClient,
+}
+
+impl BaseClient for CodeAuthSpotify {
+    fn get_http(&self) -> &HttpClient {
+        &self.http
+    }
+
+    fn get_token(&self) -> Option<&Token> {
+        self.token.as_ref()
+    }
+
+    fn get_creds(&self) -> &Credentials {
+        &self.creds
+    }
+
+    fn get_config(&self) -> &Config {
+        &self.config
+    }
+}
+
+// This could also be a macro (less important)
+impl OAuthClient for CodeAuthSpotify {
+    fn get_oauth(&self) -> &OAuth {
+        &self.oauth
+    }
 }
 
 impl CodeAuthSpotify {
@@ -17,32 +54,129 @@ impl CodeAuthSpotify {
         }
     }
 
-    pub fn prompt_for_user_token(&mut self) {
-        todo!()
-    }
-}
-
-impl BaseClient for CodeAuthSpotify {
-    fn get_http(&self) -> &HttpClient {
-        &self.http
-    }
-
-    fn get_token(&self) -> Option<&Token> {
-        self.tok.as_ref()
+    pub fn with_config(creds: Credentials, oauth: OAuth, config: Config) -> Self {
+        CodeAuthSpotify {
+            creds,
+            oauth,
+            config,
+            ..Default::default()
+        }
     }
 
-    fn get_creds(&self) -> &Credentials {
-        &self.creds
+    /// Gets the required URL to authorize the current client to start the
+    /// [Authorization Code Flow](https://developer.spotify.com/documentation/general/guides/authorization-guide/#authorization-code-flow).
+    pub fn get_authorize_url(&self, show_dialog: bool) -> ClientResult<String> {
+        let mut payload: HashMap<&str, &str> = HashMap::new();
+        let oauth = self.get_oauth();
+        let scope = oauth
+            .scope
+            .clone()
+            .into_iter()
+            .collect::<Vec<_>>()
+            .join(" ");
+        payload.insert(headers::CLIENT_ID, &self.get_creds().id);
+        payload.insert(headers::RESPONSE_TYPE, headers::RESPONSE_CODE);
+        payload.insert(headers::REDIRECT_URI, &oauth.redirect_uri);
+        payload.insert(headers::SCOPE, &scope);
+        payload.insert(headers::STATE, &oauth.state);
+
+        if show_dialog {
+            payload.insert(headers::SHOW_DIALOG, "true");
+        }
+
+        let parsed = Url::parse_with_params(auth_urls::AUTHORIZE, payload)?;
+        Ok(parsed.into_string())
     }
 
-    fn get_config(&self) -> &Config {
-        todo!()
-    }
-}
+    /// Refreshes the access token with the refresh token provided by the
+    /// [Authorization Code Flow](https://developer.spotify.com/documentation/general/guides/authorization-guide/#authorization-code-flow),
+    /// without saving it into the cache file.
+    ///
+    /// The obtained token will be saved internally.
+    #[maybe_async]
+    pub async fn refresh_user_token_without_cache(
+        &mut self,
+        refresh_token: &str,
+    ) -> ClientResult<()> {
+        let mut data = Form::new();
+        data.insert(headers::REFRESH_TOKEN, refresh_token);
+        data.insert(headers::GRANT_TYPE, headers::GRANT_REFRESH_TOKEN);
 
-// This could also be a macro (less important)
-impl OAuthClient for CodeAuthSpotify {
-    fn get_oauth(&self) -> &OAuth {
-        &self.oauth
+        let mut tok = self.fetch_access_token(&data).await?;
+        tok.refresh_token = Some(refresh_token.to_string());
+        self.token = Some(tok);
+
+        Ok(())
+    }
+
+    /// The same as `refresh_user_token_without_cache`, but saves the token
+    /// into the cache file if possible.
+    #[maybe_async]
+    pub async fn refresh_user_token(&mut self, refresh_token: &str) -> ClientResult<()> {
+        self.refresh_user_token_without_cache(refresh_token).await?;
+
+        Ok(())
+    }
+
+    /// Parse the response code in the given response url. If the URL cannot be
+    /// parsed or the `code` parameter is not present, this will return `None`.
+    pub fn parse_response_code(&self, url: &str) -> Option<String> {
+        let url = Url::parse(url).ok()?;
+        let mut params = url.query_pairs();
+        let (_, url) = params.find(|(key, _)| key == "code")?;
+        Some(url.to_string())
+    }
+
+    /// Obtains the user access token for the app with the given code without
+    /// saving it into the cache file, as part of the OAuth authentication.
+    /// The access token will be saved inside the Spotify instance.
+    #[maybe_async]
+    pub async fn request_user_token_without_cache(&mut self, code: &str) -> ClientResult<()> {
+        let oauth = self.get_oauth()?;
+        let mut data = Form::new();
+        let scopes = oauth
+            .scope
+            .clone()
+            .into_iter()
+            .collect::<Vec<_>>()
+            .join(" ");
+        data.insert(headers::GRANT_TYPE, headers::GRANT_AUTH_CODE);
+        data.insert(headers::REDIRECT_URI, oauth.redirect_uri.as_ref());
+        data.insert(headers::CODE, code);
+        data.insert(headers::SCOPE, scopes.as_ref());
+        data.insert(headers::STATE, oauth.state.as_ref());
+
+        self.token = Some(self.fetch_access_token(&data).await?);
+
+        Ok(())
+    }
+
+    /// Tries to open the authorization URL in the user's browser, and returns
+    /// the obtained code.
+    ///
+    /// Note: this method requires the `cli` feature.
+    #[cfg(feature = "cli")]
+    fn get_code_from_user(&self) -> ClientResult<String> {
+        use crate::client::ClientError;
+
+        let url = self.get_authorize_url(false)?;
+
+        match webbrowser::open(&url) {
+            Ok(_) => println!("Opened {} in your browser.", url),
+            Err(why) => eprintln!(
+                "Error when trying to open an URL in your browser: {:?}. \
+                 Please navigate here manually: {}",
+                why, url
+            ),
+        }
+
+        println!("Please enter the URL you were redirected to: ");
+        let mut input = String::new();
+        std::io::stdin().read_line(&mut input)?;
+        let code = self
+            .parse_response_code(&input)
+            .ok_or_else(|| ClientError::Cli("unable to parse the response code".to_string()))?;
+
+        Ok(code)
     }
 }
